backend/requirements.txt
fastapi~=0.109.2
openai~=1.12.0
python-dotenv~=1.0.1
alembic~=1.13.1
pgvector~=0.2.5
SQLAlchemy~=2.0.27
psycopg2